src: add timer queue to list of active handles
Make process._getActiveHandles() add the timer queue in lib/timers.js
to the list of active handles.

The timer queue is backed by a single timer handle.  Before this commit,
process._getActiveHandles() only reported the handle but that's not very
useful: it doesn't tell you what timer is keeping the event loop alive.