cluster: restore v0.10.x setupMaster() behaviour
In v0.10.x, `process.argv` and `process.execArgv` would only be
evaluated and copied into `cluster.settings` on the first call to
`cluster.setupMaster()` (either directly or via `cluster.fork()`),
allowing them to be modified as needed before initializing the
settings.

In 41b75ca the behaviour was changed so that these values are
initialized at the time of the first `require('cluster')`.

Fixes #7670.