Console.debug()
Add support for Console.debug() function, as latest versions of Chrome & Firefox has. In a logical order, debug() has a lower priority than log(), although it's only intended for developers and it's not part of the normal output of the program, due to this I output it on stderr like warn() & error().