dns: add resolveSpf
support SPF records in DNS as if they were TXT during the transition to
full support for RFC 7208 sender policy framework.

allows queries for SPF records to be made by adding resolveSpf. changes
the resolveTxt test to eliminate the assumption that a TXT record is
used for SPF mechanisms (this was never true). adds a test case for
resolveSpf, currently relying on the DNS record at digitalocean.com,
which is compliant with the new standard.