http: initialization of request options
Initialize request options in the "addRequest" method instead
of "createSocket" to makes sure calls to "getName" are
always performed with the same data and so the connections
of the pool are properly reused and released.ptions-initialization

Needed because SSLAgent use a lot of fields from the options
to build the name. Those options can be send in the SSLAgent
constructor that will be added to the request options.