cluster: tests multiple open of ports in workers
Note that test/simple/test-cluster-dgram-4.js is a simplified case, since
dgram.send() for clients does an implict .bind(0), so it could also be written
as:

  A = dgram.createSocket('udp4')
  A.send(...)
  A.close();
  ... time passes...
  B = dgram.createSocket('udp4')
  B.send(...)
  B.close();

Above was the original form, dscape/lynx kills the cluster master when run in
a worker.

IMO, dgram-4 is the most serious of this, using multiple dgram client sockets
is currently impossible with cluster.

Also note test/simple/test-cluster-net-listen-2.js, the TCP equivalent of
dgram-4, passes, because TCP close causes the cluster master to be notified,
UDP close does not.

Also, note that I use the ephemeral "flag" port `0` in the tests, the same
failure is seen with any port.