Make the fs.js module configurable
fs.js now exports the function CreateConfiguredFSObject, which takes the
configuration hash and creates the fs object with that configuration.

This way the public interface of fs.js remains the same apart from a new
function added to it. The callback-is-last principle is preserved and the
module.js can still use the optimized stat, just like any other code. All that
needs to be done is to create the configured fs object:

var throwSafeFs = fs.CreateConfiguredFSObject({ throwSafe: true });