fs: add fs.fstatExistsSync
Currently `fs.fstatSync` does throw if the file does not exist,
so current code has be be wrapped by a try/catch statement every
time.

This method returns false and does not throw when a given path
does not exist, like `fs.existsSync`.

`fs.fstatSync` was is not changed and a new method is introduced
for backwards-compability.