fix grammar in README.md
The readme used to say "The Windows and OS X installers will prompt you for the location to install to"
I have made the simple grammar fix so it now says "... prompt you for the location in which to install."
This improves readability. No other wording was changed. No source code was changed.