Update path.extname documentation
Adding an additional example to `path.extname` to demonstrate the case where the first letter of the last path component is `'.'`. This case is  an interesting one, as `path.extname('.txt')`, returns an empty string. In this case, `.txt` is a potentially valid file name with some semantic value as a text file (while arguably maintaining an extension). I agree with Node's behavior in this case, but  I think the added example provides clarity on the method's behavior.