doc: fix error about MTU and UDP payload size
The UDP payload size is not limited by the MTU. UDP will rely on IP
fragmentation to be able to send large datagrams. While it is possible
to find routers that will discard IP packets with the MF bit set, this
is quite unlikely. For example, some implementation for DNS (like BIND)
defaults now to 4096 (through EDNS0) instead of 512 which is used when
DNSSEC is present to avoid fallback over TCP.

There are other reasons to avoid using too large UDP datagrams, like
latency (for RTMP) or the fact that the receiver needs to prepare a
buffer large enough to receive the datagram without knowing its size in
advance.

I kept that payload length stuff even if it is not quite exact (with
IPv6, it is possible to use larger payload). I don't think a complete
explanation should be exposed in the documentation.