net: add pauseOnConnect option to createServer()
Currently, when a server receives a new connection, the
underlying socket handle begins reading data immediately.
This causes problems when sockets are passed between processes,
as data can be read by the first process, and thus never read
by the second process. This commit allows sockets that are
constructed with a handle to be paused initially.