build: i18n: fix issue with icu toolset dependencies
So what I did here is to make the `icutools` library
actually contain the entire ICU dependencies needed
for host-side tools. Sadly, this will build ICU twice,
but avoids conflicts between host and target side.
This all seems like a gyp bug of some sort, but without
docs for `toolsets`, who’s to say?

I removed the `icuio` library as a separate target,
because it was only used by the host-side tools.

ref: https://github.com/joyent/node/issues/7676

Update: fix for Windows