QueryString optimization: moved try-catch UNTESTED
After reading and becoming aware of V8 Optimization Killers I was
reviewing the code to extract the URL module for my own uses when
I saw this performance killer - the try-catch in a loop.  The
statement seemed simple enough to move to its own function, so
I did just that and declared it outside of the loop.

Also, at worst case scenario, this module can be **two** try-catches
deep. This may be another area of performance improvement.

I have not tested this code. If someone more fluent in using node
could do a benchmark and verify my tweak, that would be great.