http: verify http method is a valid http token
RFC 2616/7230 requires that http method conform to
the token rule. This adds a check to verify that there are
no invalid characters in the method. If there are, then
we throw, otherwise fall through and continue.

Replacement for https://github.com/joyent/node/pull/9017 based on
feedback.
Targeted at v0.10