Clear domain stack after uncaught exception
After the patch caeb677 the uncaught exception handler does not clear
the domain stack after an exception in the top level domain has been
handled. This leaves unexpected domains on the domain stack. Fix this by
clearing the domain stack also in this case.

The affects node version starting with v0.10.34, but not v0.11.