fs: fix link and symlink error messages
Currently, the error messages for missing arguments of symlink, link,
symlinkSync and linkSync are swapped. Additionally they are named
inconsistently all over the sources and the docs.

In addition, #8920 shows that the meaning of `srcpath` and `dstpath`
isn't clear and the docs aren't comprehensive enough.

This PR changes `srcpath` to `target` and `dstpath` to `path` in the docs and
the sources and make the code more consistent.