test: fix test-process-active-wraps for IPv4
test-process-active-wraps fails if the machine it runs on does not have
IPv6 enabled. This is due to the fact that Node first tries to use IPv6
and then falls back to using IPv4 if it is not enabled, as it can be
seen in lib/net.js lines 1100-1105. Even though binding to an IPv6
address fails, there is an extra handle created in the process.

This fixes the test by adding different assertions for when only IPv4
is available.