test: run v8 tests
2nd pass with the v8 tests. moves the new targets
into a separate gyp as requested. uses an alternate
build for cutest that excludes the one file that requires
c++11 (there's always that *one* test that uses auto
and ruins it for everyone else).

To use:

```
./configure --with-v8-tests
make
make test-v8
```

```
./configure --with-intl=full-icu --with-v8-tests
make
make test-v8-intl
```

A rudimentary start at support on win32 is included
but it likely needs more work.