fs: add autoClose option to fs.WriteStream
Add support to fs.createWriteStream and fs.WriteStream for an autoClose
option that behaves similarly to the autoClose option supported by
fs.createReadStream and fs.ReadStream.

When an instance of fs.WriteStream created with autoClose === true finishes,
it is not destroyed. Its underlying fd is not closed and it is the
responsibility of the user to close it.

Fixes #20880