test: fix test-process-active-wraps for IPv4
test-process-active-wraps fails if the machine it runs on does not have
IPv6 enabled. This is due to the fact that Node first tries to use IPv6
and then falls back to using IPv4 if it is not enabled, as it can be
seen in lib/net.js lines 1100-1105. Even though binding to an IPv6
address fails, there is an extra handle created in the process.

This fixes the test by checking if IPv6 is supported and then binding
either to the IPv6 loopback if it is, or the IPv4 loopback if it is
not. By binding to a specific address we avoid the fall back
behavior and the test runs consistently regardless of whether IPv6
is enabled or not.