buffer: add method readBits for reading slices of a byte
The readBits method allows for reading bits that might be inside of a byte,
most commonly, a nibble.  Bits are fairly common in the hardware world and
often times a byte can be segmented multiple times.  readBits allows you to
read the byte and deem valuable information from it.  Generally this will
only be in the form of an integer value.