test: update ssl options matrix tests
Update the ssl options matrix tests according to recent changes that
deprecate RC4 in the default ciphers list.

This change makes sure that to be able to use RC4, RC4 ciphers need to
be explicitly specified by both client and server in the options, or
--enable-legacy-cipher-list=v0.1038 needs to be passed by the side that
doesn't explicitly pass the RC4 cipher.

It also tests that the RC4-MD5 cipher suite has to be specified
explicitly by both client and server in order to be used, and that it's
not allowed by either side if passing undefined or null as the "ciphers"
option when using tls.connect or tls.Server.