configure: fix empty prefix
autoconf-based configure scripts (which this one is modeled after)
support to pass an empty prefix via --prefix=.
this is required for systems that have binaries directly in /bin
rather than /usr/bin etc (one example is sabotage linux).

the fix is easy: rather than testing options.prefix for "if var",
which treats an empty string the same as "None", we check for
"if var not None".

fixes #5197